Put comms into connection.
