+++ docs/reference/interface.md
@@ -1,4 +1,3 @@
 :::async_kernel.interface
 :::async_kernel.interface.base
-:::async_kernel.interface.zmq
 :::async_kernel.interface.ip_app
