+++ src/async_kernel/interface/base.py
@@ -8,17 +8,14 @@ import importlib.util
 import logging
 import os
 import sys
-import time
 import weakref
 from contextlib import asynccontextmanager
-from datetime import UTC, datetime
-from typing import TYPE_CHECKING, Any, Generic, Literal, Never, Self, final
-from uuid import uuid4
+from typing import TYPE_CHECKING, Any, Generic, Literal, Self, final
 
 import anyio
 from aiologic.lowlevel import AsyncLibraryNotFoundError, current_async_library
 from traitlets import import_item, traitlets
-from traitlets.config import Config, Configurable, LoggingConfigurable
+from traitlets.config import Config, Configurable
 from traitlets.config.application import Application, ClassesType
 from typing_extensions import override
 
@@ -27,22 +24,18 @@ import async_kernel.event_loop
 from async_kernel import utils
 from async_kernel.caller import Caller
 from async_kernel.common import Fixed
-from async_kernel.pending import Pending, PendingMessage, ProtectedPending
+from async_kernel.pending import PendingMessage, ProtectedPending
 from async_kernel.typing import (
     Backend,
     BuffersType,
     Channel,
     Content,
     Hosts,
-    Job,
     Message,
-    MessageProtocol,
     MsgHeader,
     MsgType,
-    MsgTypeNoReply,
     NoValue,
     RunSettings,
-    T,
     T_interface_co,
     T_shell_co,
 )
@@ -50,9 +43,10 @@ from async_kernel.typing import (
 if TYPE_CHECKING:
     from collections.abc import AsyncGenerator, Callable
 
+    from async_kernel.connection.base import Connection
     from async_kernel.kernel import Kernel
 
-__all__ = ["BaseInterface", "BaseMessage", "Connection", "HasInterface", "PendingMessage"]
+__all__ = ["HasInterface", "Interface", "PendingMessage"]
 
 
 def extract_header(msg_or_header: dict[str, Any]) -> MsgHeader | dict:
@@ -87,143 +81,7 @@ class DictValueLiteralEval(traitlets.Dict):
         return d
 
 
-class BaseMessage(LoggingConfigurable, anyio.AsyncContextManagerMixin, MessageProtocol):
-    """The base for messaging between kernel interfaces and clients."""
-
-    callers: Fixed[Self, dict[Literal[Channel.shell, Channel.control], Caller]] = Fixed[
-        Self, dict[Literal[Channel.shell, Channel.control], Caller]
-    ](dict)
-    """The callers used by the messaging application."""
-
-    started: Fixed[Self, ProtectedPending] = Fixed(ProtectedPending)
-    ""
-    stopped: Fixed[Self, ProtectedPending] = Fixed(ProtectedPending)
-    ""
-
-    session_id = Fixed(lambda _: str(uuid4()))
-    "Used to identify this object as the `session` in a message header."
-
-    bsession: Fixed[Self, bytes] = Fixed[Self, bytes](lambda c: c["owner"].session_id.encode())
-    "Used to identfiy this object as the origin of a message."
-
-    log: logging.Logger
-
-    _pending_messages: Fixed[Self, dict[str, PendingMessage[Any]]] = Fixed(dict)
-
-    @asynccontextmanager
-    async def __asynccontextmanager__(self) -> AsyncGenerator[Self]:
-        """A context manager to open the channels in a caller named 'Control'."""
-        async with Caller() as caller, caller.get(name="Control") as caller_ctrl:
-            self.callers[Channel.shell] = caller
-            self.callers[Channel.control] = caller_ctrl
-            # Open channels
-            pen_channels: Pending[None] = caller_ctrl.call_soon(self.open_channels)
-            await self.started
-            try:
-                yield self
-            finally:
-                self.stopped.set_result(None)
-                await pen_channels.wait(shield=True)
-                del pen_channels
-
-    async def open_channels(self) -> None:
-        """Open the channels, set ready when ready block until stopped."""
-        self.started.set_result(None)
-        await self.stopped
-
-    @override
-    def handle_incoming_msg(self, msg: Message, ident: list[bytes]) -> None:
-        raise NotImplementedError
-
-    @override
-    def handle_reply(self, msg: Message) -> None:
-        # Thread: undefined
-        if (parent := msg.get("parent_header")) and (f := self._pending_messages.pop(parent["msg_id"], None)):
-            self.log.debug("Received %s %s", msg["header"]["msg_type"], msg)
-            f.set_result(msg)
-
-    @property
-    def as_owner(self) -> Callable[[], Self]:
-        """Provides a callable with reference to self."""
-        return lambda: self
-
-    @override
-    def msg(
-        self,
-        msg_type: str | MsgType,
-        content: T | None = None,
-        *,
-        channel: Channel,
-        parent: Message | dict[str, Any] | None = None,
-        header: MsgHeader | dict[str, Any] | None = None,
-        metadata: dict[str, Any] | None = None,
-        buffers: BuffersType = None,
-    ) -> Message[T]:
-        parent = parent or utils.get_parent_message()
-        if header is None:
-            header = MsgHeader(
-                date=datetime.now(UTC),
-                msg_id=str(uuid4()),
-                msg_type=msg_type,
-                session=self.session_id,
-                username="",
-                version=async_kernel.kernel_protocol_version,
-            )
-        return Message(
-            channel=channel,
-            header=header,  # pyright: ignore[reportArgumentType]
-            parent_header=extract_header(parent),  # pyright: ignore[reportArgumentType]
-            content={} if content is None else content,
-            metadata=metadata if metadata is not None else {},
-            buffers=buffers,
-        )
-
-    @final
-    def _base_send_msg(self, msg: Message, ident: bytes | list[bytes] | None = None) -> Message:
-        self.transmit_msg(msg, [] if ident is None else ident if isinstance(ident, list) else [ident])
-        return msg
-
-    @override
-    @final
-    def send_message(
-        self,
-        msg: Message,
-        ident: bytes | list[bytes] | None = None,
-    ) -> PendingMessage[Content]:
-        """Sends the message to the other side (client for kernel and vice versa) and returns a PendingMessage."""
-        if MsgType(msg["header"]["msg_type"]) in MsgTypeNoReply:
-            msg_ = f"{msg['header']['msg_type']} does not send a reply! Use `send_message_no_reply` instead."
-            raise TypeError(msg_)
-        self.log.debug("Send mssage %s %s", msg["header"]["msg_type"], msg)
-        self._pending_messages[msg["header"]["msg_id"]] = pen = PendingMessage()
-        pen.metadata.update(parent=self._base_send_msg(msg, ident))
-        return pen
-
-    @override
-    @final
-    def send_message_no_reply(self, msg: Message, ident: bytes | list[bytes] | None = None) -> None:
-        self._base_send_msg(msg, ident)
-
-    @override
-    def send_reply(self, job: Job, content: dict, /, *, buffers: BuffersType = None) -> None:
-        if "status" not in content:
-            content["status"] = "ok"
-        msg = self.msg(
-            job["msg"]["header"]["msg_type"].replace("request", "reply"),
-            content=content,
-            parent=job["msg"],
-            channel=job["msg"]["channel"],
-        )
-        self.send_message_no_reply(msg, job["ident"])
-        if msg:
-            self.log.debug("send_reply %s", msg)
-
-    @override
-    def transmit_msg(self, msg: Message, ident: list[bytes]) -> None:
-        raise NotImplementedError
-
-
-class BaseInterface(Application, anyio.AsyncContextManagerMixin, Generic[T_shell_co]):
+class Interface(Application, anyio.AsyncContextManagerMixin, Generic[T_shell_co]):
     """The base class for kernel interface (singleton).
 
     The interface creates the kernel and provides external communication. It is also
@@ -253,26 +111,26 @@ class BaseInterface(Application, anyio.AsyncContextManagerMixin, Generic[T_shell
     aliases: dict[str | tuple[str, ...], str] = (  # pyright: ignore[reportIncompatibleVariableOverride]
         Application.aliases
         | {
-            ("name", "n"): "BaseInterface.kernel_name",
+            ("name", "n"): "Interface.kernel_name",
             ("f", "connection_file"): "ZMQConnection.connection_file",
-            "launcher": "BaseInterface.launcher",
+            "launcher": "Interface.launcher",
             "timeout": "BaseShell.timeout",
-            "kernel_class": "BaseInterface.kernel_class",
-            "shell_class": "BaseInterface.shell_class",
+            "kernel_class": "Interface.kernel_class",
+            "shell_class": "Interface.shell_class",
             "help_links": "Kernel.help_links",
             "supported_features": "Kernel.supported_features",
-            "interface_class": "BaseInterface.interface_class",
-            "host": "BaseInterface.host",
-            "host_options": "BaseInterface.host_options",
-            "backend_options": "BaseInterface.backend_options",
-            "backend": "BaseInterface.backend",
+            "interface_class": "Interface.interface_class",
+            "host": "Interface.host",
+            "host_options": "Interface.host_options",
+            "backend_options": "Interface.backend_options",
+            "backend": "Interface.backend",
         }
         | Application.aliases
     )
     ""
     flags = {
-        "quiet": ({"BaseInterface": {"quiet": True}}, "Only send stdout/stderr to output stream."),
-        "no-quiet": ({"BaseInterface": {"quiet": False}}, "Only send stdout/stderr to output stream."),
+        "quiet": ({"Interface": {"quiet": True}}, "Only send stdout/stderr to output stream."),
+        "no-quiet": ({"Interface": {"quiet": False}}, "Only send stdout/stderr to output stream."),
     } | Application.flags
     ""
 
@@ -291,7 +149,7 @@ class BaseInterface(Application, anyio.AsyncContextManagerMixin, Generic[T_shell
     """Options for starting the backend."""
 
     interface_class: traitlets.Type[type[Self], type[Self] | str] = traitlets.Type(
-        "async_kernel.interface.base.BaseInterface"
+        "async_kernel.interface.base.Interface"
     ).tag(  # pyright: ignore[reportAssignmentType]
         config=True
     )
@@ -373,7 +231,7 @@ class BaseInterface(Application, anyio.AsyncContextManagerMixin, Generic[T_shell
 
     @traitlets.default("autostart_connections")
     def _default_autostart_connections(self) -> list[str]:
-        return ["async_kernel.interface.zmq.ZMQConnection"] if sys.platform != "emscripten" else []
+        return ["async_kernel.connection.zmq.ZMQConnection"] if sys.platform != "emscripten" else []
 
     @traitlets.default("shell_class")
     def _default_shell_class(self):
@@ -415,7 +273,7 @@ class BaseInterface(Application, anyio.AsyncContextManagerMixin, Generic[T_shell
         **kwargs: Any,
     ) -> None:
         app = e = None
-        if BaseInterface._instance:
+        if Interface._instance:
             msg = "An interface already exists!"
             raise RuntimeError(msg)
         try:
@@ -434,10 +292,10 @@ class BaseInterface(Application, anyio.AsyncContextManagerMixin, Generic[T_shell
                 raise e
 
     def __new__(cls, argv: list | NoValue | None = NoValue, /, **kwargs) -> Self:  # noqa: ARG004  # pyright: ignore[reportInvalidTypeForm]
-        if BaseInterface._instance:
+        if Interface._instance:
             msg = "An interface already exists!"
             raise RuntimeError(msg)
-        BaseInterface._instance = inst = super().__new__(cls, **kwargs)
+        Interface._instance = inst = super().__new__(cls, **kwargs)
         return inst
 
     def __init__(
@@ -476,8 +334,8 @@ class BaseInterface(Application, anyio.AsyncContextManagerMixin, Generic[T_shell
                     self.log.addHandler(handler)
 
     def _on_stopped(self, _) -> None:
-        if BaseInterface._instance is self:
-            BaseInterface._instance = None
+        if Interface._instance is self:
+            Interface._instance = None
         self.log.info("%s, stopped", self)
 
     @override
@@ -511,7 +369,7 @@ class BaseInterface(Application, anyio.AsyncContextManagerMixin, Generic[T_shell
             - If there is an `asyncio` or `trio` event loop already running in the desired thread;
                 start asynchronously instead (`async with interface: ...`).
         """
-        if BaseInterface._instance is not self:
+        if Interface._instance is not self:
             msg = "This interface is not the global instance!"
             raise RuntimeError(msg)
 
@@ -607,6 +465,8 @@ class BaseInterface(Application, anyio.AsyncContextManagerMixin, Generic[T_shell
         self.log.info("Interface started: %s", self.summary)
         self.started.set_result(None)
         del self.iopub_send
+        from async_kernel.connection.base import Connection  # noqa: PLC0415
+
         for pth in self.autostart_connections:
             try:
                 if issubclass(cls := import_item(pth), Connection):
@@ -650,7 +510,8 @@ class BaseInterface(Application, anyio.AsyncContextManagerMixin, Generic[T_shell
                 channel=Channel.stdin,
                 # The client is assumed to have set the 'identity' of the stdin socket to 'session.bsession'.
             ),
-            ident=job["msg"]["header"]["session"].encode(),
+            # ident=job["msg"]["header"]["session"].encode(),
+            ident=job["ident"],
         )
         if current_pen := self.callers[Channel.shell].current_pending():
             current_pen.add_done_callback(lambda _: pen_reply.cancel(""))
@@ -692,7 +553,7 @@ class BaseInterface(Application, anyio.AsyncContextManagerMixin, Generic[T_shell
 
 
 class HasInterface(Generic[T_interface_co]):
-    """A mixin class providing a reference to the global [interface][async_kernel.interface.base.BaseInterface].
+    """A mixin class providing a reference to the global [interface][async_kernel.interface.base.Interface].
 
     This class is designed to be compatible with [Configurable][] objects enabling the sharing
     of configuration and log objects. The global _interface_ must exist before creating subclass
@@ -737,80 +598,13 @@ class HasInterface(Generic[T_interface_co]):
 
         # Register class for configuration
         if issubclass(cls, Configurable):
-            BaseInterface.classes.insert(0, cls)
+            Interface.classes.insert(0, cls)
 
     def __new__(cls, *args, **kwargs) -> Self:
 
-        if not (interface := BaseInterface._instance):  # pyright: ignore[reportPrivateUsage]
-            msg = "A global BaseInterface has not been created yet!"
+        if not (interface := Interface._instance):  # pyright: ignore[reportPrivateUsage]
+            msg = "A global Interface has not been created yet!"
             raise RuntimeError(msg)
         inst = new_(cls) if (new_ := super().__new__) is object.__new__ else new_(cls, *args, **kwargs)
         inst._interface = weakref.ref(interface)
         return inst
-
-
-class Connection(HasInterface[T_interface_co], BaseMessage, Generic[T_interface_co]):
-    """Provides a connection to the interface for messaging."""
-
-    def __init__(self, **kwargs: Any) -> None:
-        super().__init__(**kwargs)
-        ctx = super().__asynccontextmanager__()
-
-        def start() -> ProtectedPending[Any]:
-            # Nesting inside __init__ means this is only called once for normal usage.
-            async def run(stopped=self.stopped) -> None:
-                async with ctx:
-                    await stopped
-
-            del self.start
-            try:
-                interface = self.parent
-                assert not interface.stopping.done()
-
-                interface.started.add_done_callback(lambda _: interface.callers[Channel.shell].call_soon(run))
-                interface.stopped.add_done_callback(self.stop)
-            except Exception as e:
-                self.started.set_exception(e)
-                self.stop()
-            return self.started
-
-        self.start = start
-
-    @override
-    def __asynccontextmanager__(self) -> Never:
-        msg = f"Directly using async context of {self} is not allowed!"
-        raise RuntimeError(msg)
-
-    @override
-    async def open_channels(self) -> None:
-        """Open the channels, set ready when ready block until stopped."""
-        self.parent.refresh_connections(self)
-        self.started.set_result(None)
-        await self.stopped
-        self.parent.refresh_connections()
-
-    def start(self) -> ProtectedPending[Any]:
-        """Required to start the connection."""
-        return self.started
-
-    @override
-    def handle_incoming_msg(self, msg: Message, ident: list[bytes]) -> None:
-
-        match msg["channel"]:
-            case Channel.control | Channel.shell:
-                self.parent.kernel.handle_request(
-                    Job(msg=msg, ident=ident, received_time=time.monotonic(), owner=self.as_owner)
-                )
-            case Channel.stdin:
-                self.handle_reply(msg)
-            case _:
-                self.log.debug("Unhandled message %r %r", msg, ident)
-
-    def stop(self, _=None, /) -> ProtectedPending[Any]:
-        """Stop the connection."""
-        self.stopped.set_result(None)
-        self.parent.stopped.remove_done_callback(self.stop)
-        return self.stopped
-
-    def connection_info(self) -> str:
-        return ""
