+++ tests/test_base_interface.py
@@ -6,13 +6,14 @@ import weakref
 from typing import TYPE_CHECKING
 
 import anyio
+import anyio.lowlevel
 import pytest
 from aiologic import Event
 from traitlets.config.configurable import Configurable
 
 import async_kernel
-from async_kernel.client.base import LocalClient
-from async_kernel.interface import BaseInterface, HasInterface
+from async_kernel.connection.base import LocalClient
+from async_kernel.interface import HasInterface, Interface
 from async_kernel.shell.base import BaseShell
 from async_kernel.typing import Channel, Job, MsgType
 
@@ -26,20 +27,20 @@ class TestBaseInterface:
     async def test_instance_does_not_exist(self, anyio_backend: Backend):
 
         with pytest.raises(RuntimeError, match="An instance does not exist"):
-            BaseInterface.instance()
+            Interface.instance()
 
     async def test_instance_not_a_subclass(self, anyio_backend: Backend):
 
-        class BaseInterfaceSub(BaseInterface):
+        class BaseInterfaceSub(Interface):
             pass
 
-        async with BaseInterface():
+        async with Interface():
             with pytest.raises(TypeError, match="An instance exists but it is not an instance of"):
                 BaseInterfaceSub.instance()
 
     async def test_gc(self, anyio_backend: Backend):
         collected = Event()
-        async with BaseInterface() as interface:
+        async with Interface() as interface:
             weakref.finalize(interface, collected.set)
             ref = weakref.ref(interface)
             del interface
@@ -47,20 +48,20 @@ class TestBaseInterface:
         with anyio.move_on_after(2):
             while not collected:
                 gc.collect()
-                await anyio.sleep(0)
+                await anyio.lowlevel.checkpoint()
         if obj := ref():
             referrers = gc.get_referrers(obj)
             assert not referrers
 
     async def test_already_initialized(self, anyio_backend: Backend):
         for _ in range(3):
-            async with BaseInterface() as interface:
+            async with Interface() as interface:
                 with pytest.raises(RuntimeError, match="Already initialized!"):
                     interface.initialize()
 
     async def test_input_request_no_handler(self, anyio_backend: Backend):
 
-        async with BaseInterface(shell_class=BaseShell), LocalClient() as client:
+        async with Interface(shell_class=BaseShell), LocalClient() as client:
             msg = client.msg(MsgType.input_request, channel=Channel.stdin)
             job = Job(msg=msg, owner=client.as_owner, ident=[], received_time=time.monotonic())
             with pytest.raises(RuntimeError, match="A handler is not available"):
@@ -68,12 +69,12 @@ class TestBaseInterface:
 
     async def test_stop(self, anyio_backend: Backend) -> None:
 
-        async with BaseInterface(shell_class=BaseShell) as interface:
+        async with Interface(shell_class=BaseShell) as interface:
             interface.stop()
 
     async def test_base_shell(self, anyio_backend: Backend):
 
-        async with BaseInterface(shell_class=BaseShell) as interface:
+        async with Interface(shell_class=BaseShell) as interface:
             assert isinstance(interface.kernel.shell, BaseShell)
             assert "name:" in interface.kernel.shell.banner
             assert isinstance(interface.kernel.shell.user_ns, dict)
@@ -92,7 +93,7 @@ class TestBaseInterface:
 
     async def test_base_shell_displayhook(self, anyio_backend: Backend, mocker):
 
-        async with BaseInterface(shell_class=BaseShell) as interface:
+        async with Interface(shell_class=BaseShell) as interface:
             iopub_send = mocker.patch.object(interface, "iopub_send")
             with async_kernel.utils.show_result(True):
                 interface.kernel.shell.displayhook(123)
@@ -101,12 +102,12 @@ class TestBaseInterface:
             assert str(list(iopub_send.call_args)[1]) == expected
 
     async def test_stop_early(self, anyio_backend: Backend):
-        app = BaseInterface(shell_class=BaseShell)
+        app = Interface(shell_class=BaseShell)
         app.stop()
         with pytest.raises(RuntimeError, match="This interface is not the global instance!"):
             app.start()
         with pytest.raises(RuntimeError, match="An instance does not exist"):
-            BaseInterface.instance()
+            Interface.instance()
         with pytest.raises(RuntimeError, match="Stopped early"):
             async with app:
                 pass
@@ -115,11 +116,11 @@ class TestBaseInterface:
         def bad_loop_factory():
             raise RuntimeError
 
-        app = BaseInterface(shell_class=BaseShell, backend_options={"loop_factory": bad_loop_factory})
-        assert BaseInterface._instance is app
+        app = Interface(shell_class=BaseShell, backend_options={"loop_factory": bad_loop_factory})
+        assert Interface._instance is app
         with pytest.raises(RuntimeError):
             app.start()
-        assert BaseInterface._instance is None
+        assert Interface._instance is None
 
 
 class TestHasInterface:
